+++ src/EloquentViewableServiceProvider.php
@@ -15,6 +15,7 @@ namespace CyrildeWit\EloquentViewable;
 
 use Illuminate\Support\ServiceProvider;
 use Jaybizzle\CrawlerDetect\CrawlerDetect;
+use Illuminate\Cache\Repository as CacheRepository;
 use CyrildeWit\EloquentViewable\Resolvers\HeaderResolver;
 use CyrildeWit\EloquentViewable\Resolvers\IpAddressResolver;
 use CyrildeWit\EloquentViewable\Contracts\View as ViewContract;
@@ -31,9 +32,21 @@ class EloquentViewableServiceProvider extends ServiceProvider
      */
     public function boot()
     {
-        // $this->registerMiddleware();
-        $this->registerMigrations();
-        $this->registerPublishing();
+        if ($this->app->runningInConsole()) {
+            $config = $this->app->config['eloquent-viewable'];
+
+            $this->publishes([
+                __DIR__.'/../config/eloquent-viewable.php' => $this->app->configPath('eloquent-viewable.php'),
+            ], 'config');
+
+            if (! class_exists('CreateViewsTable')) {
+                $timestamp = date('Y_m_d_His', time());
+
+                $this->publishes([
+                    __DIR__.'/../migrations/create_views_table.php.stub' => database_path("/migrations/{$timestamp}_create_views_table.php"),
+                ], 'migrations');
+            }
+        }
     }
 
     /**
@@ -47,6 +60,12 @@ class EloquentViewableServiceProvider extends ServiceProvider
             __DIR__.'/../config/eloquent-viewable.php',
             'eloquent-viewable'
         );
+
+        $this->app->when(Views::class)
+            ->needs(CacheRepository::class)
+            ->give(function () : CacheRepository {
+                return $this->app['cache']->store(config('eloquent-viewable.cache.store'));
+            });
 
         $this->app->bind(ViewContract::class, View::class);
 
@@ -62,37 +81,5 @@ class EloquentViewableServiceProvider extends ServiceProvider
         $this->app->singleton(CrawlerDetectorContract::class, CrawlerDetectAdapter::class);
         $this->app->singleton(IpAddressResolverContract::class, IpAddressResolver::class);
         $this->app->singleton(HeaderResolverContract::class, HeaderResolver::class);
-    }
-
-    /**
-     * Register the package's migrations.
-     *
-     * @return void
-     */
-    protected function registerMigrations()
-    {
-        if ($this->app->runningInConsole()) {
-            $this->loadMigrationsFrom(__DIR__.'/../migrations');
-        }
-    }
-
-    /**
-     * Setup the resource publishing groups for Eloquent Viewable.
-     *
-     * @return void
-     */
-    protected function registerPublishing()
-    {
-        if ($this->app->runningInConsole()) {
-            $config = $this->app->config['eloquent-viewable'];
-
-            $this->publishes([
-                __DIR__.'/../config/eloquent-viewable.php' => $this->app->configPath('eloquent-viewable.php'),
-            ], 'config');
-
-            $this->publishes([
-                __DIR__.'/../migrations/2018_11_10_125700_create_views_table.php' => $this->app->databasePath('migrations/2018_11_10_125700_create_views_table.php'),
-            ], 'migrations');
-        }
     }
 }