+++ tests/Unit/ViewsTest.php
@@ -23,8 +23,8 @@ use CyrildeWit\EloquentViewable\Tests\TestCase;
 use CyrildeWit\EloquentViewable\Tests\TestHelper;
 use CyrildeWit\EloquentViewable\Tests\Stubs\Models\Post;
 use CyrildeWit\EloquentViewable\Contracts\CrawlerDetector;
-use CyrildeWit\EloquentViewable\Tests\Stubs\Models\Apartment;
 use CyrildeWit\EloquentViewable\Contracts\IpAddressResolver;
+use CyrildeWit\EloquentViewable\Tests\Stubs\Models\Apartment;
 
 class ViewsTest extends TestCase
 {
@@ -67,7 +67,7 @@ class ViewsTest extends TestCase
     }
 
     /** @test */
-    public function it_can_record_a_view_with_a_session_delay()
+    public function it_does_not_record_views_if_session_delay_is_active()
     {
         views($this->post)
             ->delayInSession(Carbon::now()->addMinutes(10))
@@ -81,24 +81,30 @@ class ViewsTest extends TestCase
     }
 
     /** @test */
-    public function it_can_record_a_view_under_a_tag()
+    public function it_can_record_a_view_with_session_delay_where_delay_is_an_integer()
     {
         views($this->post)
-            ->tag('customTag')
+            ->delayInSession(10)
             ->record();
 
-        $this->assertEquals(1, View::where('tag', 'customTag')->count());
+        views($this->post)
+            ->delayInSession(10)
+            ->record();
+
+        $this->assertEquals(1, View::count());
     }
 
-    // /** @test */
-    // public function it_can_record_a_view_under_multiple_tag()
-    // {
-    //     views($this->post)
-    //         ->tag('firstTag', 'secondTag')
-    //         ->record();
+    /** @test */
+    public function it_can_record_a_view_under_a_collection()
+    {
+        views($this->post)
+            ->collection('customCollection')
+            ->record();
 
-    //     $this->assertEquals(1, View::count());
-    // }
+        views($this->post)->record();
+
+        $this->assertEquals(1, View::where('collection', 'customCollection')->count());
+    }
 
     /** @test */
     public function it_can_record_a_view_with_a_custom_ip_address()
@@ -153,6 +159,17 @@ class ViewsTest extends TestCase
         $this->assertEquals(6, views($this->post)->period(Period::since(Carbon::parse('2018-01-10')))->count());
         $this->assertEquals(4, views($this->post)->period(Period::upto(Carbon::parse('2018-02-15')))->count());
         $this->assertEquals(4, views($this->post)->period(Period::create(Carbon::parse('2018-01-15'), Carbon::parse('2018-03-10')))->count());
+    }
+
+    /** @test */
+    public function it_can_count_the_views_with_a_collection()
+    {
+        views($this->post)->collection('custom')->record();
+        views($this->post)->collection('custom')->record();
+        views($this->post)->record();
+
+        $this->assertEquals(2, views($this->post)->collection('custom')->count());
+        $this->assertEquals(1, views($this->post)->count());
     }
 
     /** @test */