+++ CHANGELOG.md
@@ -5,7 +5,43 @@ All notable changes to `Eloquent Viewable` will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased v3.0.0]
+## Unreleased
+
+## [v3.2.0] (2019-03-03)
+
+### Added
+
+- Add support for Laravel 5.8
+
+### Changed
+
+- Use String and Array classes instead of the helper functions
+
+## [v3.1.0] (2019-01-29)
+
+### Fixed
+
+- Fixed the ability to pass an integer to the `delayInSession` method without getting an error
+- Type cast the cached views count otherwise PHP's type hint will fail
+
+### Added
+
+- Added the ability to override the visitor's unique ID that's used to distinguish unique views
+- Added the ability to specify a cache store that should be used by this package
+
+## [v3.0.2] (2018-12-25)
+
+### Fixed
+
+- The method `delayInSession` isn't working properly
+
+## [v3.0.1] (2018-12-25)
+
+### Fixed
+
+- Publishing package migrations results in error (#133)
+
+## [v3.0.0] (2018-12-17)
 
 ### Added
 
@@ -14,10 +50,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added `HeaderResolver` contract with implementation
 - Added `VisitorCookieRepository` class
 - Added global helper `views`
+- Added `collection` column to views table shema
+- Added `withinPeriod` scope to `View` model
+- Added `uniqueVisitor` scope to `View` model
 
 ### Changed
 
-- The public views method on viewable models will now return an instance of the `Views` class
+- Bumped minimum required PHP version to ^7.1
+- Require viewable models to implement the `Viewable` contract
+- Added global `views()` helper
 - Remove IP address as fallback for visitor cookie when it doesn't exists
 - Changed the `isBot` method name to `isCrawler` in `CrawlerDetector` contract and updated the `CrawlerDetectAdapter`
 - Changed the visibility of the `$detector` property from `protected` to `private`
@@ -33,8 +74,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Removed the `ViewableService` class
 - Removed the `ProcessView` job
 - Removed the `update_views_table` migration file from `resources/database/migrations`
+- Removed `illuminate/bus` as dependency
+- Removed `illuminate/queue` as dependency
+- Removed `illuminate/routing` as dependency
 
 ## [Unreleased]
+
+## [v2.5.0] (2018-12-03)
+
+### Fixed
+
+- `orderByViewsCount` scope doesn't adhere to connection prefix
 
 ## [v2.4.3] (2018-10-21)
 
@@ -110,7 +160,12 @@ This major version contains some serious breaking changes! See the [upgrade guid
 - Removed the `addPageViewThatExpiresAt` method from the `Viewable` trait
 - The DateTransformer functionality has been removed
 
-[Unreleased]: https://github.com/cyrildewit/eloquent-viewable/compare/v2.4.0...HEAD
+[Unreleased]: https://github.com/cyrildewit/eloquent-viewable/compare/v3.2.0...HEAD
+[v3.2.0]: https://github.com/cyrildewit/eloquent-viewable/compare/v3.1.0...v3.2.0
+[v3.1.0]: https://github.com/cyrildewit/eloquent-viewable/compare/v3.0.2...v3.1.0
+[v3.0.2]: https://github.com/cyrildewit/eloquent-viewable/compare/v3.0.1...v3.0.2
+[v3.0.1]: https://github.com/cyrildewit/eloquent-viewable/compare/v3.0.0...v3.0.1
+[v3.0.0]: https://github.com/cyrildewit/eloquent-viewable/compare/v2.3.3...v3.0.0
 [v2.4.3]: https://github.com/cyrildewit/eloquent-viewable/compare/v2.3.2...v2.4.3
 [v2.4.2]: https://github.com/cyrildewit/eloquent-viewable/compare/v2.3.1...v2.4.2
 [v2.4.1]: https://github.com/cyrildewit/eloquent-viewable/compare/v2.4.0...v2.4.1